LPD-43135 Remove useless abstraction
