+++ modules/apps/static/portal-osgi-web/portal-osgi-web-servlet-jsp-compiler/src/main/java/com/liferay/portal/osgi/web/servlet/jsp/compiler/internal/JspCompiler.java
@@ -201,7 +201,7 @@ public class JspCompiler {
 			}
 
 			_javaFileObjectResolvers.add(
-				new JspJavaFileObjectResolver(
+				new JavaFileObjectResolver(
 					bundleWiring, _jspBundleWiring, _bundleWiringPackageNames));
 		}
